Moved out CheckBountyRespawnAndAward into a separate file
