+++ src/GUI/GUIBountyHunting.h
@@ -1,0 +1,4 @@
+#pragma once
+
+
+void CheckBountyRespawnAndAward();
