+++ src/GUI/GUIWindow.h
@@ -370,7 +370,6 @@ char sub_4637E0_is_there_popup_onscreen();
 void ClickNPCTopic(DIALOGUE_TYPE topic);
 void _4B3FE5_training_dialogue(int a4);
 void OracleDialogue();
-void CheckBountyRespawnAndAward();
 std::string _4B254D_SkillMasteryTeacher(int trainerInfo);
 std::string BuildDialogueString(const char *lpsz, uint8_t uPlayerID,
                            struct ItemGen *a3, int eventId, int a5,
